requirements.txt
# Network Rail STOMP client - Latest stable version
stomp.py==8.0.1

# Timezone handling - Latest stable version
pytz==2023.3.post1

# Kafka client - Latest stable version (if needed for future use)
confluent-kafka==2.3.0

# Logging and utilities - Latest stable version
colorama==0.4.6

# Optional: For future web dashboard features
# flask==3.0.0
# requests==2.31.0
# plotly==5.17.0
# pandas==2.1.4
